Implement A-MoreOOP
