+++ src/main/java/Event.java
@@ -1,5 +1,4 @@
 public class Event extends Task {
-
     protected String at;
 
     public Event(String description, String at) {
