+++ src/main/java/Todo.java
@@ -1,5 +1,4 @@
 public class Todo extends Task {
-
     public Todo(String description) {
         super(description);
     }
